refactor: split prompts into separate files
